alterado para suportar Bradesco pag_for
