+++ l10n_br_account_banking_payment_cnab/febraban/pag_for/__init__.py
@@ -20,43 +20,3 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-
-
-from ..cnab_240 import Cnab240
-import re
-import string
-from decimal import *
-
-
-class Bradesco240(Cnab240):
-
-    def __init__(self):
-        super(Cnab240, self).__init__()
-        from cnab240.bancos import bradesco
-        self.bank = bradesco
-
-    def _prepare_header(self):
-        """
-
-        :param order:
-        :return:
-        """
-        vals = super(Bradesco240, self)._prepare_header()
-        return vals
-
-    def _prepare_segmento(self, line):
-        """
-
-        :param line:
-        :return:
-        """
-        vals = super(Bradesco240, self)._prepare_segmento(line)
-        return vals
-
-    # Override cnab_240.nosso_numero. Diferentes números de dígitos entre CEF e Itau
-    def nosso_numero(self, format):
-        digito = format[-1:]
-        carteira = format[:3]
-        nosso_numero = re.sub(
-            '[%s]' % re.escape(string.punctuation), '', format[3:-1] or '')
-        return carteira, nosso_numero, digito
